refactor(frontend): split public-csv-download god-object (S5 of #141)
Pure code movement — no behavior changes. PublicCsvDownload drops
from 1143 to 717 LOC; two new focused collaborators take the
extracted slices.

  - CsvDownloadValidator (291 LOC) — validate_form_access,
    validate_hash_only, validate_cpf_requirement, plus the private
    record_download_log_entry helper they share.
  - CsvDownloadFormInfoBuilder (277 LOC) — build_form_info plus its
    private build_restrictions_info / build_datetime_info /
    build_geolocation_info / format_locations_for_info /
    build_quiz_info helpers.

The facade keeps its public API intact so external callers
(PublicCsvExporter, FormEditorPublicCsvDownloadMetabox, Frontend) and
the test (`$this->handler->validate_cpf_requirement(...)`) keep
working unchanged. Constructor wires two private properties
(`$validator`, `$form_info_builder`); the three validation methods
become one-line forwarders; `ajax_info()` calls
`$this->form_info_builder->build_form_info(...)` inline.

Constants stay on PublicCsvDownload (SHORTCODE, ACTION,
NONCE_ACTION, META_*, DOWNLOAD_LOG_MAX). Methods inside the new
classes reference them as fully-qualified `PublicCsvDownload::*`.
The static helpers `get_audit_log_summary`, `maybe_wipe_legacy_logs`,
and the private `decrypt_log_entry_cpf` stay on the facade — they
are not validation/info-builder logic.

Test update: two `\ReflectionMethod( PublicCsvDownload::class, ... )`
tests targeting the moved `build_datetime_info` private method now
reflect on `CsvDownloadFormInfoBuilder::class` — same coverage, new
home.

Verification:
  - php -l: clean on all 3 files.
  - vendor/bin/phpunit --filter PublicCsvDownload: 42/42 OK (119 assertions).
  - vendor/bin/phpunit: 3870/3870 OK.
  - vendor/bin/phpstan analyse includes/frontend/: no errors.
